Reorganize Mods.D2k
