+++ OpenRA.Mods.D2k/Traits/Render/WithDeliveryOverlay.cs
@@ -14,7 +14,7 @@ using OpenRA.Mods.Common.Traits.Render;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.D2k.Render
+namespace OpenRA.Mods.D2k.Traits
 {
 	[Desc("Rendered when ProductionAirdrop is in progress.")]
 	public class WithDeliveryOverlayInfo : ITraitInfo, Requires<RenderSpritesInfo>, Requires<IBodyOrientationInfo>
