+++ OpenRA.Mods.D2k/Traits/World/D2kResourceLayer.cs
@@ -14,7 +14,7 @@ using System.Linq;
 using OpenRA.Mods.RA;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.D2k
+namespace OpenRA.Mods.D2k.Traits
 {
 	[Desc("Used to render spice with round borders.")]
 	public class D2kResourceLayerInfo : TraitInfo<D2kResourceLayer> { }
