+++ OpenRA.Mods.D2k/Traits/World/PaletteFromR8.cs
@@ -13,7 +13,7 @@ using OpenRA.FileSystem;
 using OpenRA.Graphics;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.D2k
+namespace OpenRA.Mods.D2k.Traits
 {
 	class PaletteFromR8Info : ITraitInfo
 	{
